create chart package split styles
