+++ pyseas/doc/TestChartStyle.py
@@ -17,6 +17,7 @@
 import pyseas.maps as psm
 import numpy as np
 import matplotlib.pyplot as plt
+from pyseas import styles
 import pyseas
 
 # This magic function attempts to reload all of PySeas, which is
@@ -30,18 +31,7 @@ pyseas._reload()
 # except for turning off clipping, which I haven't found a style
 # for.
 
-with psm.context(psm.styles.chart_style):
+with psm.context(psm.styles.charts.light):
     plt.plot(np.arange(10), np.arange(10), '--.', label='line')
     plt.xlim(0, 8)
     plt.legend()
-
-# %%
-# Note that styles is currently located in pyseas and aliased in pyseas.maps.
-# It might make sense to have pyseas incorporate styles.maps and styles.charts
-# So pyseas.styles.maps is basically the same (minus the chart style which now one uses)
-# and pyseas.styles.charts gives the charts styles (so we could have multiple ones)
-
-import pyseas
-pyseas.styles.dark
-
-# %%
